+++ TechStore/Models/Categorium.cs
@@ -5,7 +5,7 @@ namespace TechStore.Models;
 
 public partial class Categorium
 {
-    public decimal IdCategoria { get; set; }
+    public int IdCategoria { get; set; }
 
     public string Nombre { get; set; } = null!;
