+++ TechStore/Models/DetalleCompra.cs
@@ -5,17 +5,17 @@ namespace TechStore.Models;
 
 public partial class DetalleCompra
 {
-    public decimal IdDetalleCompra { get; set; }
+    public int IdDetalleCompra { get; set; }
 
-    public decimal Cantidad { get; set; }
+    public int Cantidad { get; set; }
 
     public decimal PrecioCompra { get; set; }
 
     public decimal Subtotal { get; set; }
 
-    public decimal IdCompra { get; set; }
+    public int IdCompra { get; set; }
 
-    public decimal IdProducto { get; set; }
+    public int IdProducto { get; set; }
 
     public virtual CompraProveedor IdCompraNavigation { get; set; } = null!;
