+++ TechStore/Models/HistorialInventario.cs
@@ -5,17 +5,17 @@ namespace TechStore.Models;
 
 public partial class HistorialInventario
 {
-    public decimal IdMovimiento { get; set; }
+    public int IdMovimiento { get; set; }
 
     public string TipoMovimiento { get; set; } = null!;
 
-    public decimal Cantidad { get; set; }
+    public int Cantidad { get; set; }
 
     public DateTime Fecha { get; set; }
 
     public string? Observacion { get; set; }
 
-    public decimal IdProducto { get; set; }
+    public int IdProducto { get; set; }
 
     public virtual Producto IdProductoNavigation { get; set; } = null!;
 }
