+++ TechStore/Models/Rol.cs
@@ -5,7 +5,7 @@ namespace TechStore.Models;
 
 public partial class Rol
 {
-    public decimal IdRol { get; set; }
+    public int IdRol { get; set; }
 
     public string Nombre { get; set; } = null!;
