+++ TechStore/Models/VisHistorialInventario.cs
@@ -5,13 +5,13 @@ namespace TechStore.Models;
 
 public partial class VisHistorialInventario
 {
-    public decimal IdMovimiento { get; set; }
+    public int IdMovimiento { get; set; }
 
     public string Producto { get; set; } = null!;
 
     public string TipoMovimiento { get; set; } = null!;
 
-    public decimal Cantidad { get; set; }
+    public int Cantidad { get; set; }
 
     public DateTime Fecha { get; set; }
